+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - **Major** - Added: Data Visualizer/Editor now contains a visual representation of the nodes in the area.
 - **Major** - Added: The Logic Database can now have descriptions for nodes.
 - Added: Data Editor can now edit area names.
+- Added: Data Editor can now view and edit resources.
 
 ### Metroid Prime